fix ~ skip 'verify' tests when key retrieval fails
.# Discussion

The keyserver network is somewhat unreliable and can fail to deliver requested keys
(eg, <https://github.com/docker-library/official-images/issues/4252> @@
<http://archive.is/DFnZN>). These keyserver failures lead to intermittent test failures
for 'verify' under both AppVeyor and Travis CI testing.

This changeset skips instead of failing 'verify' tests during key retrieval failures.

A more complete solution might be to move to use a sequence of keyservers, moving down
the list if a keyserver is unavailable or doesn't respond correctly. But it's not clear
whether that process should be in the library or the 'verify' testing.

Notably, this solution conflates keyserver failures (unresponsive server, broken response
[eg, "keyserver helper general error"], etc.) and failure due to an actually missing
public key (eg, "not found" or "bad public key"). Since the keys used for testing are
expected to be published and synchronized on the keyserver network, it seems ok not to
separate the two types of failure.

So, for now, this code serves as a work-around, allowing CI builds to pass/fail based on
the module code and not the flaky availability of a keyserver key retrieval.